+++ CHANGELOG.rst
@@ -8,6 +8,10 @@ Features:
 
 - Add ``skip_on_field_errors`` parameter to ``validates_schema`` (:issue:`323`). Thanks :user:`jjvattamattom` for the suggestion and :user:`d-sutherland` for the PR.
 
+Bug fixes:
+
+- Fix ``FormattedString`` serialization (:issue:`348`). Thanks :user:`acaird` for reporting.
+
 2.3.0 (2015-11-22)
 ++++++++++++++++++